+++ flopy4/mf6/component.py
@@ -6,7 +6,6 @@ from modflow_devtools.dfn import Dfn, Var
 from xattree import xattree
 
 from flopy4.io import IOMethod, Loader, Writer
-
 from flopy4.mf6.spec import fields_dict
 
 COMPONENTS = {}
